solved a bug cuasing a crash when trying to edit a task and cleaned the encryption scheme methods
