+++ app/src/main/java/com/github/se/studybuddies/ui/todo/EditToDo.kt
@@ -1,5 +1,6 @@
 package com.github.se.studybuddies.ui.todo
 
+//noinspection UsingMaterialAndMaterial3Libraries
 import android.util.Log
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
